+++ lib/pages/create/complete_habit_dialog.dart
@@ -12,7 +12,7 @@ import 'package:lotti/themes/theme.dart';
 import 'package:lotti/utils/platform.dart';
 import 'package:lotti/widgets/charts/utils.dart';
 import 'package:lotti/widgets/dashboards/dashboard_widget.dart';
-import 'package:lotti/widgets/misc/datetime_bottom_sheet.dart';
+import 'package:lotti/widgets/date_time/datetime_field.dart';
 import 'package:tinycolor2/tinycolor2.dart';
 import 'package:url_launcher/url_launcher.dart';